Nodes are created at the right place
In the function AssignBlueprint within the class TerrainPartObject, the
node positions are now assigned based upon the actual node positions
within the blueprint. Prior to this, the AssignBlueprint function either
used the same calculation as the BlueprintPart class to work out where
the nodes go (which is redundant) or placed them at an arbitrary point
(which is incorrect). Instead now, the BlueprintPart class has a new
function added called GetNodePosition. This allows AssignBlueprint to
read the actual node positions and move the newly created nodes (and
controls) accordingly.