Terrain segment length is correct at creation
Modified AssignBlueprint in TerrainPartObject to call this.Regenerate()
at the end of the function rather than calling
this.terrainPart.Regenerate() at the beginning of the function. First of
all, calling it at the end is necessary because it relies upon the
nodes, which are created within this function. Secondly,
this.Regenerate() is called instead of this.terrainPart.Regenerate(),
because it calculates the segment lengths before regenerating the
terrain.