Vertex nodes can be joined, and snap by default
Nodes now snap together by defaulted, and once snapped together they
will stay together when moved. They can be unsnapped by pressing the key
for inputNodeSelectIndividual (KeyCode.LeftAlt). Snapping can be
disabled by holding inputNodeSnapNode (KeyCode.LeftShift). Whether
snapping occurs by default or not can be set with snapByDefault, and the
snapping distance with snapMinDist. The moving of multiple nodes is
achieved by the following: the first node to be selected detects all
other nodes at its exact same position. They are then placed into an
array, and are moved along with the first selected node. Although the
term "node" is used, this only applies to the vertex nodes, and not to
the control nodes. Control nodes do not snap.