Rewrote how terrains are created
The nodes for terrains are no longer passed through the terrain
constructor. Instead, the constructor has no arguments. The terrain
piece is first created, and then can be queried for e.g. how many nodes
are required. The nodes are then passed through the new function
SetNodePositions. This change was necessary for the code that loads
levels from file, since it needed to know how many Vector2s to read from
the file before creating the terrain piece (which was impossible). This
method ensures that the information for how many nodes are required for
each terrain is centralised to within the specific terrain classes,
making any later edits or additions much simpler.