Can save and load levels
Saving and loading is fully implemented. Also, a new class was added to
aid with creating terrains: TerrainPartMaker. It is simply a matter of
creating a new instance of TerrainPartMaker with the appropriate
BlueprintPartType, adding the required nodes and calling
CreateTerrain(). This way the code for creating terrain is centralised,
rather than redundantly implementing the same terrain code in many
places.