Nodes only move as mouse moves
The offset between a mouse and a node is remembered when the node is
first selected. This same offset is applied back when the node is moved
with the mouse. In this way the node only moves when the mouse moves
(whereas before, upon first clicking on a node, the node would jump to
land so that its middle was exactly where the mouse was).