Removed redundant TerrainPartObject prefab
TerrainPartObejct prefab was no longer needed because the object
switched to be created through code in the last commit, rather than
through loading a prefab. Also, a small bug was fixed where the terrain
would turn red when making the desired segment length longer. This was
due to code left over from testing the colour code.