Moved node deactivation code
The node deactivation code was moved to EditorController.cs. This way a
bug was also easily fixed. When loading a level, nodes were not
deactivated by default. By moving this code, after loading a level, all
nodes are activated and then deactivated. This ensures that all nodes
are deactivated when a level is loaded.