Nodes snap properly
The nodes now snap based upon the distance between the center of the
two nodes. Before it snapped between the node and mouse position. This
was not a good method because it is more intuitive visually to snap
when the two node centers are nearby. The fix is to add back the mouse
offset when calculating the distance. You cannot simply use
this.GetVertexPosition() because the result of the following code ends
up modifying that position, and so it would end up locking the vertex
in place once it snaps.