Added rollers
This addition required many big changes all over the place. The term
"terrain" has changed to "ground", while "terrain" is now the general
term for something such as "ground" or "rollers". As such, GroundObject
was deleted and replaced with a base class, TerrainBase, and two
children, TerrainGround and TerrainRoller. Within TerrainGenerator.cs,
the enums were changed and added to. In this way, there are now separate
enums for ground terrain and roller terrain. The stage has also been set
for being able to use different sprites for terrains.