TerrainObjectMaker now intakes TerrainInfo
Due to the larger amount of information required to be passed through
when generating a new terrain, this data was encapsulated within a class
in TerrainGenerator called TerrainInfo. The existing TerrainInfo within
EditorInterfaceKeyboard was renamed InterfaceTerrainInfo. A function was
then added into EditorInterfaceKeyboard to convert the
InterfaceTerrainInfo data into TerrainInfo data which is then passed to
TerrainObjectMaker.