Speed now actually works
Previously the speed setting actually adjusted the torque of the
rollers, but they all ran at the same speed. This is because of the
setting maxAngularVelocity. Now, maxAngularVelocity is set according to
the desired speed, and all the torque values are the same. Also, the
direction of the wheel is stored in a separate value, called direction.