View and change roller settings in editor
Previously once a roller was created, its settings could not be changed
or viewed. Now if you click on a roller, you can view its raw settings.
If you press/hold space, its settings are overwritten with the current
settings and it is regenerated. Additionally, the speed and direction
field of TerrainRoller have been updated so that when writing, only the
speed has to be written to. For example: a signed value of -50 is
written to speed. Speed then takes on the value of 50 (because speed is
unsigned) and sets direction to -1 (because direction stores the sign).
When writing the speed to file for saving, the two are combined again.